Reorganized components and folders
