+++ src/components/Card/AlbumCard/CoverImage.jsx
